Fix for exception

+++ django_site_queue/management/commands/queue_manager_json.py
@@ -102,25 +102,26 @@ class Command(BaseCommand):
                     jsondb.save_queue_session(sitesession_file,sitesession)
                     sitesession = jsondb.get_queue_session(sitesession_file)
                     print (sitesession_file)
-                    if sitesession["status"]  == "Active":
-                        print ("Migrate to Active folder")
-                        try:
-                            LOCK_PATH = str(sitesession_file)+".lock" 
-                            lock = FileLock(LOCK_PATH)
-                            with lock:
-                                session_filename = os.path.basename(sitesession_file)
+                    if sitesession:
+                        if sitesession["status"]  == "Active":
+                            print ("Migrate to Active folder")
+                            try:
+                                LOCK_PATH = str(sitesession_file)+".lock" 
+                                lock = FileLock(LOCK_PATH)
+                                with lock:
+                                    session_filename = os.path.basename(sitesession_file)
 
-                                active_sitesession_file = "db/json/queue_sessions/active/{}/{}".format(queue_group_name,session_filename)
-                                shutil.copyfile(sitesession_file, active_sitesession_file)  
-                                time.sleep(.1)
-                                os.remove(sitesession_file)   
-                                print ("Removing file "+str(sitesession_file))
-                            try:       
-                                os.remove(LOCK_PATH)
-                            except Exception as k:
-                                print ("Error Removing "+str(LOCK_PATH))
-                                print (k)
+                                    active_sitesession_file = "db/json/queue_sessions/active/{}/{}".format(queue_group_name,session_filename)
+                                    shutil.copyfile(sitesession_file, active_sitesession_file)  
+                                    time.sleep(.1)
+                                    os.remove(sitesession_file)   
+                                    print ("Removing file "+str(sitesession_file))
+                                try:       
+                                    os.remove(LOCK_PATH)
+                                except Exception as k:
+                                    print ("Error Removing "+str(LOCK_PATH))
+                                    print (k)
 
-                        except Exception as e:
-                            print ("Error Saving File:"+str(sitesession_file))
-                            print (e)                                
+                            except Exception as e:
+                                print ("Error Saving File:"+str(sitesession_file))
+                                print (e)
